Update Project 3 .py

+++ Project 3 .py
@@ -2,9 +2,9 @@
 # accept principal , rate , time and print simple interest and amount
 
 # storing the Principal , Time and Rate  in the variables P , T and R respectively
-P = int(input("Enter the principle amount: "))
-T = int(input("Enter the time in years: "))
-R = int(input("Enter the Rate of Interest per annum"))
+P = float(input("Enter the principle amount: "))
+T = float(input("Enter the time in years: "))
+R = float(input("Enter the Rate of Interest per annum"))
 # calculating the SI
 SI = (P * R * T) / 100
 # calculation the amount
